backend/requirements.txt
asgiref==3.11.0
bcrypt==5.0.0
cachetools==5.5.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.1
Faker==40.1.2
google-auth==2.25.0
idna==3.11
pillow==12.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
